requirements.txt
anyio==4.10.0
arrow==1.3.0
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
Pygments==2.19.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.5
rich==14.1.0
six==1.17.0
sniffio==1.3.1
speedtest-cli==2.1.3
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==4.0.2
